backend/requirements.txt
fastapi
uvicorn
pydantic
libsql-client
python-dotenv
httpx
google-auth-oauthlib
google-api-python-client
twilio
